img_text_speech/requirements.txt
pytesseract==0.3.7
Pillow
pyttsx3==2.71
google_trans_new==1.1.9